Move config to config.rs
